backend/requirements.txt
azure-core
azure-search-documents
azure-identity
python-dotenv
requests
fastapi
uvicorn
openai